+++ src/main/scala/com/github/nikdon/telepooz/engine/ApiRequestExecutor.scala
@@ -45,7 +45,11 @@ abstract class ApiRequestExecutor(implicit system: ActorSystem, materializer: Ma
 
   override def apply[A](fa: RawRequest[A]): Future[A] = fa match {
     case m: GetMe.type             ⇒ go(m.name, m.payload)
+    case m: GetWebhookInfo.type    ⇒ go(m.name, m.payload)
     case m: SendMessage            ⇒ go(m.name, m.payload)
+    case m: SendGame               ⇒ go(m.name, m.payload)
+    case m: SetGameScore           ⇒ go(m.name, m.payload)
+    case m: GetGameHighScores      ⇒ go(m.name, m.payload)
     case m: ForwardMessage         ⇒ go(m.name, m.payload)
     case m: GetUpdates             ⇒ go(m.name, m.payload)
     case m: SendPhoto              ⇒ go(m.name, m.payload)