+++ src/main/scala/com/github/nikdon/telepooz/model/InlineKeyboardButton.scala
@@ -11,8 +11,15 @@ package com.github.nikdon.telepooz.model
   *                             open that chat and insert the bot‘s username and the specified inline
   *                             query in the input field. Can be empty, in which case just the bot’s username
   *                             will be inserted.
+  * @param switch_inline_query_current_chat If set, pressing the button will insert the bot‘s username and the specified
+  *                                         inline query in the current chat's input field. Can be empty, in which case
+  *                                         only the bot’s username will be inserted.
+  * @param callback_game        Description of the game that will be launched when the user presses the button.
+  *                             NOTE: This type of button must always be the first button in the first row.
   */
 case class InlineKeyboardButton(text: String,
-                                url: Option[String],
-                                callback_data: Option[String],
-                                switch_inline_query: Option[String])
+                                url: Option[String] = None,
+                                callback_data: Option[String] = None,
+                                switch_inline_query: Option[String] = None,
+                                switch_inline_query_current_chat: Option[String] = None,
+                                callback_game: Option[CallbackGame] = None)